+++ tests/specs/exportAssignment/functionReference.d.ts
@@ -1,0 +1,3 @@
+export = hello;
+
+declare function hello(): string;